Finished the Generic Multicast 1 spec.
We finished our work in writing the specification. Some fixes were
needed in the primitives. We added the properties, which are working
fine now.

This version is doing all the steps, which means that we include the
Reliable Multicast primitive. Some minor changes were necessary to only
consume messages from Atomic Broadcast when an entry exists in the
Memory.