Generic Multicast 1 (#2)
We finished our work in writing the specification. Some fixes were
needed in the primitives. We added the properties, which are working
fine now.

This version is doing all the steps, we already removed the use of
the Reliable Multicast primitive. We replaced it with the Atomic
Broadcast use.

## Shuffle the message sequences.

Shuffling the messages sequences that we use to initialize the Atomic
Broadcast primitive. We have that groups now always start with messages
delivered in different orders between groups.